3단계 구간 추가 기능 (#798)
* refactor: 한줄에 한점만 찍도록 수정

* fix: 전체 테스트 코드 수행시 오류가 발생하던 문제 수정

* feat: 3단계 요구사항 및 to-do list 추가

* feat: 지하철 구간 등록 인수테스트 추가

* refactor: sql 이 아닌 entityManager 를 통해 데이터를 초기화 하도록 수정

* feat: 지하철 구간 추가 기능 추가

* feat: 누락된 to-do list 갱신

* refactor: Station 값을 꺼내서 비교하는게 아닌 메시지를 전달해서 비교하도록 수정

* refactor: 전역 예외 형태로 변경

* refactor: 다형성을 위해 LinkedHashSet 타입 선언을 Set 으로 변경

* refactor: Line 조회시 Station jpa proxy 로 인한 fetch 조인 적용

* refactor: LineService Station 을 조회 하는 메소드 중복 제거

* refactor: 구간 재배열 메소드명을 rearrange 로 변경

* refactor: Section 의 isStandard 로 시작하는 메소드명 isEquals 로 변경

* feat: 상행/하행 역이 같은 경우는 추가 될수 없도록 유효성 검사 추가

* refactor: Sections 에서 Section 을 포함하는지 판별식을 메소드로 추출

* refactor: 디미터 법칙에 맞춰 Section 에서 distance 값을 꺼내 비교 하지않고 객체끼리 비교하도록 수정

+++ src/main/java/nextstep/subway/generic/domain/distance/DistanceConverter.java
@@ -4,15 +4,15 @@ import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
 
 @Converter(autoApply = true)
-public class DistanceConverter implements AttributeConverter<Distance, Long> {
+public class DistanceConverter implements AttributeConverter<Distance, Integer> {
 
     @Override
-    public Long convertToDatabaseColumn(final Distance attribute) {
+    public Integer convertToDatabaseColumn(final Distance attribute) {
         return attribute.getValue();
     }
 
     @Override
-    public Distance convertToEntityAttribute(final Long dbData) {
+    public Distance convertToEntityAttribute(final Integer dbData) {
         return Distance.valueOf(dbData);
     }
 }
